Navigation to artiicle added
This commit represents my first attempt to display an article in full.
The navigational elements are in place: The user can click “< Today’s
Articles” to return to the list OR they can swipe right to dismiss the
article.

The article’s body is shown in a UIWebView. The implementation is not
ideal. The UIWebView should generally take up the entire screen, but
here it cannot because the title, image, author and date are not HTML
elements - only the body contains pure HTML. This is a challenge. A
full article as HTML would be far easier to display.