Nav with UIWebView of article
This version replaces the tableVIew of an article with a full page
UIWebView. This means I’m ignoring the “body” content of the JSON
payload and using the “id” to simply load the full article.

The problems with this implementation are that:
a. the MindBodyGreen search bar/widget on the webpage behaves poorly
b. we will need to completely disable the new App security feature so
that a MBG article can link to any website which might use any other
website to store the images. You’ll notice I have NOT done that
therefore our ads and links to outside website don’t currently work
c. MindBodyGreen articles have a floating share menu at the bottom of
the page (which works nicely when shown), which make me think adding <
and > buttons to a bar across the bottom which look weird. That leaves
only a bit of space on the navigation bar at the top. Basically I need
to talk to the UX person at this point rather than cram things in.